requirements.txt
altair==5.5.0
fpdf==1.7.2
langchain==0.3.22
langchain_community==0.3.20
langchain_core==0.3.50
matplotlib==3.8.4
mongoengine==0.29.1
motor==3.7.0
networkx==3.2.1
numpy==2.2.4
openai==1.70.0
pandas==2.2.3
passlib==1.7.4
pgmpy==0.1.26
plotly==5.22.0
plotly==5.24.1
pydantic==2.11.1
pymongo==4.11.3
python-dotenv==1.1.0
python_bcrypt==0.3.2
Requests==2.32.3
scikit_learn==1.4.2
seaborn==0.13.2
streamlit==1.32.0
streamlit==1.39.0
wordcloud==1.9.4
json
mongoengine
pymongo
langchain
fpdf
pandas
numpy
